requirements.txt.txt
Flask==2.0.1
pyTelegramBotAPI==4.7.0
google-api-python-client==2.34.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
python-dotenv==0.19.0
